Moved ConfigUpdateListener and ConfigUpdateListenerRegistration to public level.
