+++ firebase-config/src/main/java/com/google/firebase/remoteconfig/internal/ConfigRealtimeHttpClient.java
@@ -1,5 +1,6 @@
 package com.google.firebase.remoteconfig.internal;
 
+import com.google.firebase.remoteconfig.FirebaseRemoteConfig;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -10,7 +11,7 @@ import java.util.Map;
  */
 public class ConfigRealtimeHttpClient {
 
-  private final Map<Integer, ConfigUpdateListener> listeners;
+  private final Map<Integer, FirebaseRemoteConfig.ConfigUpdateListener> listeners;
   private int listenerCount;
 
   public ConfigRealtimeHttpClient() {
@@ -24,11 +25,11 @@ public class ConfigRealtimeHttpClient {
   // Pauses Http stream listening
   private void pauseRealtimeStream() {}
 
-  public ConfigUpdateListenerRegistration addRealtimeConfigUpdateListener(
-      ConfigUpdateListener configUpdateListener) {
+  public FirebaseRemoteConfig.ConfigUpdateListenerRegistration addRealtimeConfigUpdateListener(
+      FirebaseRemoteConfig.ConfigUpdateListener configUpdateListener) {
     listeners.put(listenerCount, configUpdateListener);
     beginRealtimeStream();
-    return new ConfigUpdateListenerRegistration(this, listenerCount++);
+    return new FirebaseRemoteConfig.ConfigUpdateListenerRegistration(this, listenerCount++);
   }
 
   public void removeRealtimeConfigUpdateListener(int listenerKey) {
@@ -36,27 +37,5 @@ public class ConfigRealtimeHttpClient {
     if (listeners.isEmpty()) {
       pauseRealtimeStream();
     }
-  }
-
-  public static class ConfigUpdateListenerRegistration {
-    private final ConfigRealtimeHttpClient client;
-    private final int listenerKey;
-
-    public ConfigUpdateListenerRegistration(ConfigRealtimeHttpClient client, int listenerKey) {
-      this.client = client;
-      this.listenerKey = listenerKey;
-    }
-
-    public void remove() {
-      client.removeRealtimeConfigUpdateListener(listenerKey);
-    }
-  }
-
-  // Event Listener interface to be used by developers.
-  public interface ConfigUpdateListener {
-    // Call back for when Realtime fetches.
-    void onEvent();
-
-    void onError(Exception error);
   }
 }
